Adding class toggle and close buttons
Added buttons to html
Added handlers for adding the 'done' class and the button closure.
Adding the handlers both to the existing list items, and to new ones.
Added the closure button to new items.
Added (not required) preventing of removing item from the list if it has not been marked as done.