Use event delegation for performance
Changed 'mark as done' and close button handlers to be on the parent UL, meaning not as many listeners are required.
Added checking so the action is only performed for the correct element type.
Added css so the UL isn't wide, causing click events for the list items even when the space around them are clicked.